requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
sqlalchemy==2.0.25
pandas==2.2.0
numpy==1.26.3
streamlit==1.31.0
plotly==5.18.0
streamlit-autorefresh==1.0.1
apscheduler==3.10.4
openai==2.30.0
python-dotenv==1.0.0
schedule==1.2.1
